Move test file into performance test project
