+++ src/DynamoCore/Properties/AssemblyInfo.cs
@@ -28,3 +28,4 @@ using System.Runtime.CompilerServices;
 [assembly: InternalsVisibleTo("Dynamo.Storage")]
 [assembly: InternalsVisibleTo("LibraryViewExtension")]
 [assembly: InternalsVisibleTo("ViewExtensionLibraryTests")]
+[assembly: InternalsVisibleTo("DynamoPerformanceTests")]
